crawler/requirements.txt
requests==2.31.0
PyGithub==2.1.1
python-dotenv==1.0.0
schedule==1.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.2
